+++ src/widgets/NavigationBar/Notifications.tsx
